move eventlogging to own module, start distinguishing ll events
